# docs/openapi-core.yaml
openapi: 3.1.0
info:
  title: MediConnect Core API
  version: 2026-03-14
  description: |
    Core contract for web and mobile clients. This spec is intentionally compact and
    focused on shared post-login modules.
servers:
  - url: https://{host}
    variables:
      host:
        default: app.mediconnect.local
components:
  securitySchemes:
    SupabaseBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
  /api/auth/mobile/session:
    get:
      summary: Check mobile session state
      responses:
        "200":
          description: Session state
    post:
      summary: Bind mobile access token to server cookie session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                accessToken:
                  type: string
                deviceId:
                  type: string
      responses:
        "200":
          description: Session bound
    delete:
      summary: Clear mobile session
      responses:
        "200":
          description: Session cleared
...